backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pdfplumber==0.11.0
PyPDF2==3.0.1
deep-translator==1.11.4
gTTS==2.5.1
deepl==1.17.0
Werkzeug==3.0.1
